simplify code for broadcast changes
