+++ lib/realtime/application.ex
@@ -65,8 +65,6 @@ defmodule Realtime.Application do
         Realtime.Latency,
         {Registry, keys: :duplicate, name: Realtime.Registry},
         {Registry, keys: :unique, name: Realtime.Registry.Unique},
-        {Registry, keys: :unique, name: Realtime.BroadcastChanges.Handler.Registry},
-        {Registry, keys: :unique, name: Realtime.BroadcastChanges.Listener.Registry},
         {Task.Supervisor, name: Realtime.TaskSupervisor},
         {PartitionSupervisor,
          child_spec: DynamicSupervisor,
@@ -79,10 +77,6 @@ defmodule Realtime.Application do
          max_restarts: 5},
         {DynamicSupervisor,
          name: Realtime.Tenants.Migrations.DynamicSupervisor, strategy: :one_for_one},
-        {PartitionSupervisor,
-         child_spec: DynamicSupervisor,
-         strategy: :one_for_one,
-         name: Realtime.BroadcastChanges.Listener.DynamicSupervisor},
         {PartitionSupervisor,
          child_spec: DynamicSupervisor,
          strategy: :one_for_one,
